refactor(runtime): consolidate micro-files by domain in console and terminal
Fold same-domain CLI types, HTTP infra, system-font routes, and Theater
CRUD into their owning modules. Drop the styles barrel and inline the
unused agent duration helper. Keep check-report.ts so update-check tests
can mock checkUpdateStatus at a module boundary.
