--- runtime/fleet-console/cli/styles/index.ts
@@ -1,18 +1,0 @@
-export {
-  ANSI_BOLD,
-  ANSI_DIM,
-  ANSI_RESET,
-  ASCII_FLEET_BANNER,
-  FLEET_ACCENT,
-  FLEET_COMMAND,
-  FLEET_OPTION,
-  GRADIENT_COLORS,
-  command,
-  dim,
-  option,
-  paint,
-  resolveColorEnabled,
-  section,
-  stripAnsi,
-  type ResolveColorEnabledOptions,
-} from "./tokens.js";
