--- runtime/fleet-console/cli/update/types.ts
@@ -1,4 +1,0 @@
-export interface UpdateCommandIo {
-  readonly stdout: Pick<NodeJS.WriteStream, "write">;
-  readonly stderr: Pick<NodeJS.WriteStream, "write">;
-}
