+++ runtime/fleet-console/core/host/http-infra.ts
@@ -15,3 +15,32 @@ export const CONSOLE_SECURITY_HEADERS = {
 export function withSecurityHeaders(headers: http.OutgoingHttpHeaders = {}): http.OutgoingHttpHeaders {
   return { ...CONSOLE_SECURITY_HEADERS, ...headers };
 }
+
+export const SSE_KEEPALIVE_INTERVAL_MS = 30_000;
+
+export function encodeSseData(event: string, data: unknown): string {
+  return `event: ${event}\ndata: ${JSON.stringify(data)}\n\n`;
+}
+
+export function startSseKeepaliveLifecycle(
+  res: http.ServerResponse,
+  onCleanup: () => void,
+): () => void {
+  res.setTimeout(0);
+  const interval = setInterval(() => {
+    if (res.writableEnded || res.destroyed) return;
+    res.write(": keepalive\n\n");
+  }, SSE_KEEPALIVE_INTERVAL_MS);
+  interval.unref();
+
+  let cleanedUp = false;
+  const cleanup = () => {
+    if (cleanedUp) return;
+    cleanedUp = true;
+    clearInterval(interval);
+    onCleanup();
+  };
+  res.on("close", cleanup);
+  res.on("error", cleanup);
+  return cleanup;
+}
